+++ Challenger.Api/Controllers/GymRecordController.cs
@@ -13,20 +13,24 @@ namespace Challenger.Api.Controllers
     public class GymRecordController : Controller
     {
         private readonly IGymRecordRepository _gymRecordRepository;
+        private readonly IUserRepository _userRepository;
         private IMapper _mapper;
 
         public GymRecordController(
             IGymRecordRepository gymRecordRepository,
+            IUserRepository userRepository,
             IMapper mapper)
         {
             _gymRecordRepository = gymRecordRepository;
+            _userRepository = userRepository;
             _mapper = mapper;
         }
 
         [HttpGet]
         public async Task<GymRecordDto[]> Get()
         {
-            var all = await _gymRecordRepository.GetAll();
+            var userId = await _userRepository.GetIdByEmail(User.Identity.Name);
+            var all = await _gymRecordRepository.GetAllForUser(userId);
             return _mapper.Map<GymRecordDto[]>(all);
         }
 
@@ -34,7 +38,7 @@ namespace Challenger.Api.Controllers
         public async Task<GymRecordDto> Add([FromBody] GymRecordDto record)
         {
             var entity = _mapper.Map<GymRecord>(record);
-            entity.UserId = 1;
+            entity.UserId = await _userRepository.GetIdByEmail(User.Identity.Name);
             _gymRecordRepository.Add(entity);
             await _gymRecordRepository.SaveChanges();
 
@@ -44,6 +48,12 @@ namespace Challenger.Api.Controllers
         [HttpPatch]
         public async Task<JsonResult> Update([FromBody] GymRecordDto record)
         {
+            var userId = await _userRepository.GetIdByEmail(User.Identity.Name);
+            if (userId != record.UserId)
+            {
+                return Json(new { IsSuccess = false });
+            }
+
             await _gymRecordRepository.Update(_mapper.Map<GymRecord>(record));
             await _gymRecordRepository.SaveChanges();
 
@@ -54,6 +64,12 @@ namespace Challenger.Api.Controllers
         public async Task<JsonResult> Delete(long id)
         {
             var toDelete = await _gymRecordRepository.Get(id);
+            var userId = await _userRepository.GetIdByEmail(User.Identity.Name);
+            if (userId != toDelete.UserId)
+            {
+                return Json(new { IsSuccess = false });
+            }
+
             _gymRecordRepository.Remove(toDelete);
             await _gymRecordRepository.SaveChanges();
