+++ Challenger.Api/Controllers/MeasurementController.cs
@@ -14,20 +14,24 @@ namespace Challenger.Api.Controllers
     public class MeasurementController : Controller
     {
         private readonly IMeasurementRepository _measurementRepository;
+        private readonly IUserRepository _userRepository;
         private readonly IMapper _mapper;
 
         public MeasurementController(
             IMeasurementRepository measurementRepository,
+            IUserRepository userRepository,
             IMapper mapper)
         {
             _measurementRepository = measurementRepository;
+            _userRepository = userRepository;
             _mapper = mapper;
         }
 
         [HttpGet]
         public async Task<MeasurementDto[]> Get()
         {
-            var all = await _measurementRepository.GetAll();
+            var userId = await _userRepository.GetIdByEmail(User.Identity.Name);
+            var all = await _measurementRepository.GetAllForUser(userId);
             return _mapper.Map<MeasurementDto[]>(all);
         }
 
@@ -35,7 +39,7 @@ namespace Challenger.Api.Controllers
         public async Task<MeasurementDto> Add([FromBody] MeasurementDto measurement)
         {
             var entity = _mapper.Map<Measurement>(measurement);
-            entity.UserId = 1;
+            entity.UserId = await _userRepository.GetIdByEmail(User.Identity.Name);
             _measurementRepository.Add(entity);
             await _measurementRepository.SaveChanges();
 
@@ -45,6 +49,12 @@ namespace Challenger.Api.Controllers
         [HttpPatch]
         public async Task<JsonResult> Update([FromBody] MeasurementDto measurement)
         {
+            var userId = await _userRepository.GetIdByEmail(User.Identity.Name);
+            if (userId != measurement.UserId)
+            {
+                return Json(new { IsSuccess = false });
+            }
+
             await _measurementRepository.Update(_mapper.Map<Measurement>(measurement));
             await _measurementRepository.SaveChanges();
 
@@ -55,6 +65,12 @@ namespace Challenger.Api.Controllers
         public async Task<JsonResult> Delete(long id)
         {
             var toDelete = await _measurementRepository.Get(id);
+            var userId = await _userRepository.GetIdByEmail(User.Identity.Name);
+            if (userId != toDelete.UserId)
+            {
+                return Json(new { IsSuccess = false });
+            }
+
             _measurementRepository.Remove(toDelete);
             await _measurementRepository.SaveChanges();
